requirements.txt
PyYAML==6.0.3
paho-mqtt==2.1.0